+++ pymemcache/test/test_client.py
@@ -797,6 +797,23 @@ class TestClient(ClientTestMixin, unittest.TestCase):
         with pytest.raises(MemcacheIllegalInputError):
             _set()
 
+    def test_set_key_with_noninteger_expire(self):
+        client = self.make_client([b''])
+
+        class _OneLike(object):
+            """object that looks similar to the int 1"""
+            def __str__(self):
+                return "1"
+
+        for noreply in (True, False):
+            for expire in (1.5, _OneLike(), "1"):
+                def _set():
+                    client.set(b'finekey', b'finevalue',
+                               noreply=noreply, expire=expire)
+
+                with pytest.raises(MemcacheIllegalInputError):
+                    _set()
+
     def test_set_many_socket_handling(self):
         client = self.make_client([b'STORED\r\n'])
         result = client.set_many({b'key': b'value'}, noreply=False)