+++ ChangeLog.rst
@@ -9,6 +9,12 @@ New in version 2.2.1
 --------------------
 * Fix ``flags`` when setting multiple differently-typed values at once.
 
+New in version 3.0.0 (unreleased)
+---------------------------------
+
+* Validate integer inputs for ``expire``, ``delay``, ``incr``, ``decr``, and
+  ``memlimit`` -- non-integer values now cause ``MemcacheIllegalInputError``s
+
 New in version 2.2.0
 --------------------
 * Drop official support for Python 3.4.